+++ extensions/pokedex/src/ability.tsx
@@ -5,7 +5,7 @@ import abilities from "./statics/abilities.json";
 
 const generations = groupBy(abilities, "generation");
 
-export default function Ability() {
+export default function PokeAbilities() {
   return (
     <List throttle isShowingDetail={true}>
       {Object.entries(generations).map(([generation, abilities]) => {