+++ extensions/pokedex/src/pokemon.tsx
@@ -6,26 +6,51 @@ import {
   getPreferenceValues,
 } from "@raycast/api";
 import groupBy from "lodash.groupby";
-import { useMemo, useState } from "react";
-import PokemonDetail from "./components/detail";
+import orderBy from "lodash.orderby";
+import shuffle from "lodash.shuffle";
+import { useEffect, useState } from "react";
+import PokeProfile from "./components/profile";
 import TypeDropdown from "./components/type_dropdown";
-
 import pokedex from "./statics/pokedex.json";
+import { getOfficialArtworkImg, getPixelArtImg, localeName } from "./utils";
 
-const { language } = getPreferenceValues();
+const { language, artwork } = getPreferenceValues();
+let columns: number;
 
-export default function SearchPokemon() {
+let getContent: (id: number, formId?: number) => string;
+switch (artwork) {
+  case "pixel":
+    getContent = getPixelArtImg;
+    columns = 6;
+    break;
+  default:
+    getContent = getOfficialArtworkImg;
+    break;
+}
+
+export default function NationalPokedex() {
   const [type, setType] = useState<string>("all");
+  const [sort, setSort] = useState<string>("lowest");
+  const [randomization, setRandomization] = useState<boolean>(false);
+  const [pokemons, setPokemons] = useState(pokedex);
 
-  const pokemons = useMemo(() => {
-    return type != "all"
-      ? pokedex.filter((p) => p.types.includes(type))
-      : pokedex;
-  }, [type]);
+  useEffect(() => {
+    const shuffled = shuffle(pokemons);
+    setPokemons(shuffled);
+  }, [randomization]);
+
+  useEffect(() => {
+    const sorted = orderBy(pokedex, ...sort.split("|"));
+    const filtered =
+      type != "all" ? sorted.filter((p) => p.types.includes(type)) : sorted;
+
+    setPokemons(filtered);
+  }, [type, sort]);
 
   return (
     <Grid
       throttle
+      columns={columns}
       searchBarPlaceholder="Search for Pokémon by name or Pokédex number"
       searchBarAccessory={
         <TypeDropdown type="grid" command="Pokémon" onSelectType={setType} />
@@ -39,17 +64,48 @@ export default function SearchPokemon() {
                 return (
                   <Grid.Item
                     key={pokemon.id}
-                    content={pokemon.artwork}
-                    title={language === "1" ? pokemon.jp_name : pokemon.name}
+                    content={getContent(pokemon.id)}
+                    title={localeName(pokemon, language)}
                     subtitle={`#${pokemon.id.toString().padStart(4, "0")}`}
                     keywords={[pokemon.id.toString(), pokemon.name]}
                     actions={
                       <ActionPanel>
-                        <Action.Push
-                          title="Show Details"
-                          icon={Icon.Sidebar}
-                          target={<PokemonDetail id={pokemon.id} />}
-                        />
+                        <ActionPanel.Section title="Information">
+                          <Action.Push
+                            title="Pokémon Profile"
+                            icon={Icon.Sidebar}
+                            target={<PokeProfile id={pokemon.id} />}
+                          />
+                        </ActionPanel.Section>
+                        <ActionPanel.Section title="Randomize">
+                          <Action
+                            title="Surprise Me!"
+                            icon={Icon.Shuffle}
+                            onAction={() => setRandomization(!randomization)}
+                          />
+                        </ActionPanel.Section>
+                        <ActionPanel.Section title="Sort By">
+                          <Action
+                            title="Number (Lowest First)"
+                            icon={Icon.ArrowUp}
+                            onAction={() => setSort("id|asc")}
+                          />
+                          <Action
+                            title="Number (Highest First)"
+                            icon={Icon.ArrowDown}
+                            onAction={() => setSort("id|desc")}
+                          />
+                          <Action
+                            title="Name (A-Z)"
+                            icon={Icon.Text}
+                            onAction={() => setSort("name|asc")}
+                          />
+                          <Action
+                            title="Name (Z-A)"
+                            icon={Icon.Text}
+                            onAction={() => setSort("name|desc")}
+                          />
+                        </ActionPanel.Section>
                       </ActionPanel>
                     }
                   />